add deletion validation

+++ core/src/main/java/ru/stand/contest/restapi_yandex/service/ItemService.java
@@ -6,24 +6,20 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.stand.contest.restapi_yandex.dto.SystemItemImportDto;
 import ru.stand.contest.restapi_yandex.dto.SystemItemImportRequest;
 import ru.stand.contest.restapi_yandex.entity.Item;
+import ru.stand.contest.restapi_yandex.handler.ItemNotFoundException;
 import ru.stand.contest.restapi_yandex.mapper.ItemMapper;
 import ru.stand.contest.restapi_yandex.model.SystemItem;
 import ru.stand.contest.restapi_yandex.repository.ItemRepository;
 import ru.stand.contest.restapi_yandex.validator.ImportsValidator;
 
-import javax.transaction.Transactional;
 import java.sql.Date;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.TreeSet;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -44,7 +40,8 @@ public class ItemService {
         for (SystemItemImportRequest request : systemItem) {
             itemsList.addAll(request.getItems().stream()
                     .map(item -> ItemMapper.INSTANCE.toEntity(item, new Date(request.getUpdateDate().getTime())))
-                    .peek(item -> checkValidItem.validateItem(item))
+                    //.peek(item -> checkValidItem.validateItem(item))
+                    .peek(itemRepository::save)
                     .collect(Collectors.toList()));
         }
 
@@ -61,9 +58,14 @@ public class ItemService {
     }
 
 
+    @Transactional
     public ResponseEntity<Void> deleteById(String id, java.util.Date date) {
-        itemRepository.deleteById(UUID.fromString(id));
-        return ResponseEntity.ok().build();
+        Optional<Item> optionalItem = itemRepository.findById(UUID.fromString(id));
+        if (optionalItem.isPresent()) {
+            itemRepository.deleteById(UUID.fromString(id));
+            return ResponseEntity.ok().build();
+        }
+        throw new ItemNotFoundException("Элемент не найден.");
     }
 
 }
